refactor: move test to define config
