+++ tests/configure.spec.ts
@@ -9,33 +9,9 @@
 
 import { test } from '@japa/runner'
 import { FileSystem } from '@japa/file-system'
-import { IgnitorFactory } from '@adonisjs/core/factories'
 import Configure from '@adonisjs/core/commands/configure'
 
-import { BASE_URL } from '../tests_helpers/index.js'
-
-async function setupApp() {
-  const ignitor = new IgnitorFactory()
-    .withCoreProviders()
-    .withCoreConfig()
-    .create(BASE_URL, {
-      importer: (filePath) => {
-        if (filePath.startsWith('./') || filePath.startsWith('../')) {
-          return import(new URL(filePath, BASE_URL).href)
-        }
-
-        return import(filePath)
-      },
-    })
-
-  const app = ignitor.createApp('web')
-  await app.init().then(() => app.boot())
-
-  const ace = await app.container.make('ace')
-  ace.ui.switchMode('raw')
-
-  return { ace, app }
-}
+import { setupApp } from '../tests_helpers/index.js'
 
 async function setupFakeAdonisproject(fs: FileSystem) {
   await Promise.all([
